chore: remove unused deps and fix some clippy::pedantic

+++ metalink-downloader/src/cli.rs
@@ -12,11 +12,11 @@ pub struct Cli {
 pub enum Commands {
     /// Download file
     DownloadFile {
-        /// url to download
+        /// `url` to download
         #[arg(short, long)]
         url: url::Url,
 
-        /// target_dir to download
+        /// `target_dir` to download
         #[arg(short, long)]
         target_dir: PathBuf,
 
@@ -25,8 +25,8 @@ pub enum Commands {
         user_agent: String,
 
         /// Max number of download threads to use
-        #[arg(long, default_value_t=2, value_parser = clap::value_parser!(u64).range(2..))]
-        max_threads: u64,
+        #[arg(long, default_value_t=2, value_parser = clap::value_parser!(u16).range(2..))]
+        max_threads: u16,
     },
 
     /// Dryrun the planning phase
@@ -55,7 +55,7 @@ pub enum Commands {
         user_agent: String,
 
         /// Max number of download threads to use per file
-        #[arg(long, default_value_t=2, value_parser = clap::value_parser!(u64).range(2..))]
-        max_threads_per_file: u64,
+        #[arg(long, default_value_t=2, value_parser = clap::value_parser!(u16).range(2..))]
+        max_threads_per_file: u16,
     },
 }
